refactor: Name을 도메인에 맞게 구체적인 이름으로 바꾸고, 이름 검증 로직을 Name 안에 구현
