--- src/main/java/vo/Name.java
@@ -1,43 +1,0 @@
-package vo;
-
-import java.util.Objects;
-
-public class Name {
-    private final String name;
-
-    private Name(String name) {
-        this.name = name;
-    }
-
-    public static Name of(String name) {
-        return new Name(name);
-    }
-
-    public String getValue() {
-        return name;
-    }
-
-    public int length() {
-        return name.length();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Name name1 = (Name) o;
-        return name.equals(name1.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-
-    @Override
-    public String toString() {
-        return "Name{" +
-                "name='" + name + '\'' +
-                '}';
-    }
-}
